training code

+++ scripts/vis_single.py
@@ -4,6 +4,7 @@ import sys
 import time
 import trimesh
 import torch
+import glob
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '../'))
 
@@ -30,14 +31,11 @@ rotation = int(args.rotation)
 
 dataset = save_folder.split("/")[-1].split("_")[0]
 
-format = 'obj'
-scale = 180.0
-
-mesh_file = os.path.join(f'../data/{dataset}/scans/{subject}', f'{subject}.{format}')
+mesh_file = os.path.join(f'../data/{dataset}/scans/{subject}', f'{subject}.obj')
 fit_file = f'../data/{dataset}/fits/{subject}/smplx_param.pkl'
     
 rescale_fitted_body, _ = load_fit_body(fit_file, 
-                                        scale, 
+                                        180.0, 
                                         smpl_type='smplx', 
                                         smpl_gender='neutral')
 
@@ -54,19 +52,21 @@ for y in range(0, 360, 360//rotation):
     
     os.makedirs(os.path.dirname(vis_file), exist_ok=True)
     
-    calib = load_calib(calib_file).cuda()
-    calib_verts = projection(smpl_verts, calib, format='tensor')
-    (xy,z) = calib_verts.split([2,1],dim=1)
-    smpl_vis = get_visibility(xy, z, smpl_faces)
+    if not os.path.exists(vis_file):
     
-    if args.mode == 'debug':
-        mesh = trimesh.Trimesh(smpl_verts.cpu().numpy(), smpl_faces.cpu().numpy(), process=False)
-        mesh.visual.vertex_colors = torch.tile(smpl_vis, (1,3)).numpy()
-        mesh.export(vis_file.replace("pt", "obj"))
-    
-    torch.save(smpl_vis, vis_file)
+        calib = load_calib(calib_file).cuda()
+        calib_verts = projection(smpl_verts, calib, format='tensor')
+        (xy,z) = calib_verts.split([2,1],dim=1)
+        smpl_vis = get_visibility(xy, z, smpl_faces)
         
-done_jobs = len(os.listdir(save_folder))
+        if args.mode == 'debug':
+            mesh = trimesh.Trimesh(smpl_verts.cpu().numpy(), smpl_faces.cpu().numpy(), process=False)
+            mesh.visual.vertex_colors = torch.tile(smpl_vis, (1,3)).numpy()
+            mesh.export(vis_file.replace("pt", "obj"))
+        
+        torch.save(smpl_vis, vis_file)
+        
+done_jobs = len(glob.glob(f"{save_folder}/*/vis"))
 all_jobs = len(os.listdir(f"../data/{dataset}/scans"))
 print(
     f"Finish visibility computing {subject}| {done_jobs}/{all_jobs} | Time: {(time.time()-t0):.0f} secs")
